feat: upgrade React app from v17 to v19
- React 17 → 19 with createRoot API
- Migrated from Create React App to Vite
- Converted App class component to functional component
- Removed unused React imports (new JSX transform)
- Renamed .js → .jsx for proper JSX handling
- Updated Redux (v5), React-Redux (v9), React Router DOM (v7)
- Updated sass, prettier, and dev tooling
- Fixed Modal component with proper useRef pattern
- Build verified
