+++ react-app/package.json
@@ -3,53 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "vite",
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "format": "prettier --write \"src/**/*.{js,jsx}\"",
-    "lint": "eslint \"src/**/*.{js,jsx}\" --quiet",
     "fastify-dev": "npm --prefix ../fastify-api-server start",
     "start-react-fastify": "concurrently \"npm run fastify-dev\" \"VITE_API=http://0.0.0.0:3000/api npm run dev\"",
     "start-react-func-swa": "npx @azure/static-web-apps-cli@latest start http://localhost:3000 --api-location ../api --run \"npm run start\"",
     "start-react-fastify-swa": "concurrently \"npm run fastify-dev\" \"npx @azure/static-web-apps-cli@latest start http://0.0.0.0:3000 --api-devserver-url http://0.0.0.0:3000 --host=0.0.0.0 --run 'VITE_API=/api npm run start -- --host 0.0.0.0'\""
   },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "proxy": "http://localhost:7071/",
   "engines": {
     "node": ">=20.0.0"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.3",
     "axios": "^1.13.5",
-    "bulma": "^0.9.2",
-    "history": "^5.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^6.30.3",
-    "react-scripts": "^5.0.1",
-    "redux": "^4.0.4",
-    "redux-saga": "^1.0.5",
-    "redux-thunk": "^2.4.2",
-    "sass": "^1.43.4"
+    "bulma": "^0.9.4",
+    "react": "^19.2.5",
+    "react-dom": "^19.2.5",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^7.14.2",
+    "redux": "^5.0.1",
+    "redux-saga": "^1.0.5"
   },
   "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "^7.23.1",
-    "prettier": "^2.2.1"
+    "@vitejs/plugin-react": "^6.0.1",
+    "prettier": "^3.8.3",
+    "sass": "^1.99.0",
+    "vite": "^8.0.10"
   }
 }
