+++ react-app/src/components/AuthLogin.jsx
@@ -1,16 +1,16 @@
-import React from 'react';
+export function AuthLogin(props) {
+  const { provider } = props;
 
-export function AuthLogout() {
   function goAuth() {
     const { pathname } = window.location;
-    const redirect = `post_logout_redirect_uri=${pathname}`;
-    const url = `/.auth/logout?${redirect}`;
+    const redirect = `post_login_redirect_uri=${pathname}`;
+    const url = `/.auth/login/${provider}?${redirect}`;
     window.location.href = url;
   }
 
   return (
     <div className="auth-link" onClick={goAuth} role="button" tabIndex="0">
-      Logout
+      {provider}
     </div>
   );
 }
