+++ react-app/src/components/ButtonFooter.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const ButtonFooter = ({
   label,
   className,
