+++ react-app/src/components/ListHeader.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { NavLink } from 'react-router-dom';
 
 const ListHeader = ({
