+++ react-app/src/components/NavBar.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { NavLink } from 'react-router-dom';
 import { AuthLogin } from './AuthLogin';
 import { AuthLogout } from './AuthLogout';
