+++ react-app/src/index.jsx
@@ -1,15 +1,12 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import { applyMiddleware, compose, createStore } from 'redux';
+import { applyMiddleware, compose, legacy_createStore as createStore } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 import App from './App';
 import './index.css';
-import * as serviceWorker from './serviceWorker';
 import app, { productSaga, discountSaga } from './store';
 
-// create and configure reduxer middleware ( saga is a middleware )
 const sagaMiddleware = createSagaMiddleware();
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
@@ -21,17 +18,11 @@ const store = createStore(
 sagaMiddleware.run(productSaga);
 sagaMiddleware.run(discountSaga);
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <Provider store={store}>
     <BrowserRouter>
       <App />
     </BrowserRouter>
   </Provider>,
-
-  document.getElementById('root'),
 );
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
