+++ react-app/src/products/ProductDetail.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { ButtonFooter, InputDetail } from '../components';
