+++ react-app/src/products/ProductList.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { ButtonFooter, CardContent } from '../components';
