+++ react-app/src/store/config.js
@@ -1,3 +1,3 @@
-const API = process.env.REACT_APP_API || '/api';
+const API = import.meta.env.VITE_API || '/api';
 
 export { API as default };
